+++ lib/jest-helpers.js
@@ -6,6 +6,7 @@ const cheerio = require('cheerio')
 const nunjucks = require('nunjucks')
 const outdent = require('outdent')
 const yaml = require('js-yaml')
+const { configureAxe } = require('jest-axe')
 
 const sass = require('node-sass')
 const sassRender = util.promisify(sass.render)
@@ -182,4 +183,15 @@ function htmlWithClassName ($, className) {
   return $.html($component)
 }
 
-module.exports = { render, renderHtml, renderAndInitialise, getExamples, htmlWithClassName, renderSass, renderTemplate }
+/**
+ * As we're testing incomplete HTML fragments, we don't expect there to be a
+ * skip link, or for them to be contained within landmarks.
+ */
+const axe = configureAxe({
+  rules: {
+    'skip-link': { enabled: false },
+    region: { enabled: false }
+  }
+})
+
+module.exports = { axe, render, renderHtml, renderAndInitialise, getExamples, htmlWithClassName, renderSass, renderTemplate }
