+++ src/govuk/components/tabs/template.test.js
@@ -2,9 +2,7 @@
  * @jest-environment jsdom
  */
 
-const axe = require('../../../../lib/axe-helper')
-
-const { render, getExamples } = require('../../../../lib/jest-helpers')
+const { axe, render, getExamples } = require('../../../../lib/jest-helpers')
 
 const examples = getExamples('tabs')
